+++ code/renderer/tr_shade_calc.c
@@ -23,6 +23,7 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 
 #include "tr_local.h"
 // -EC-: avoid using ri.ftol
+// NOLINTNEXTLINE(bugprone-macro-parentheses) — `table` is always a static float[] array, never a complex expression
 #define	WAVEVALUE( table, base, amplitude, phase, freq )  ((base) + table[ (int64_t)( ( ( (phase) + tess.shaderTime * (freq) ) * FUNCTABLE_SIZE ) ) & FUNCTABLE_MASK ] * (amplitude))
 
 static float *TableForFunc( genFunc_t func )
@@ -325,7 +326,7 @@ static void DeformText( const char *text ) {
 
 	// draw each character
 	for ( i = 0 ; i < len ; i++ ) {
-		ch = text[i];
+		ch = (byte)text[i];
 		ch &= 255;
 
 		if ( ch != ' ' ) {
@@ -756,6 +757,7 @@ void RB_CalcModulateColorsByFog( unsigned char *colors ) {
 	RB_CalcFogTexCoords( texCoords[0] );
 
 	for ( i = 0; i < tess.numVertexes; i++, colors += 4 ) {
+		// NOLINTNEXTLINE(clang-analyzer-core.CallAndMessage) — RB_CalcFogTexCoords initializes texCoords; analyzer doesn't model the per-vertex write
 		float f = 1.0 - R_FogFactor( texCoords[i][0], texCoords[i][1] );
 		colors[0] *= f;
 		colors[1] *= f;
@@ -777,6 +779,7 @@ void RB_CalcModulateAlphasByFog( unsigned char *colors ) {
 	RB_CalcFogTexCoords( texCoords[0] );
 
 	for ( i = 0; i < tess.numVertexes; i++, colors += 4 ) {
+		// NOLINTNEXTLINE(clang-analyzer-core.CallAndMessage) — RB_CalcFogTexCoords initializes texCoords; analyzer doesn't model the per-vertex write
 		float f = 1.0 - R_FogFactor( texCoords[i][0], texCoords[i][1] );
 		colors[3] *= f;
 	}
@@ -796,6 +799,7 @@ void RB_CalcModulateRGBAsByFog( unsigned char *colors ) {
 	RB_CalcFogTexCoords( texCoords[0] );
 
 	for ( i = 0; i < tess.numVertexes; i++, colors += 4 ) {
+		// NOLINTNEXTLINE(clang-analyzer-core.CallAndMessage) — RB_CalcFogTexCoords initializes texCoords; analyzer doesn't model the per-vertex write
 		float f = 1.0 - R_FogFactor( texCoords[i][0], texCoords[i][1] );
 		colors[0] *= f;
 		colors[1] *= f;
