+++ code/renderervk/tr_shade_calc.c
@@ -23,6 +23,7 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 
 #include "tr_local.h"
 // -EC-: avoid using ri.ftol
+// NOLINTNEXTLINE(bugprone-macro-parentheses) — `table` is always a static float[] array, never a complex expression
 #define	WAVEVALUE( table, base, amplitude, phase, freq )  ((base) + table[ (int64_t)( ( ( (phase) + tess.shaderTime * (freq) ) * FUNCTABLE_SIZE ) ) & FUNCTABLE_MASK ] * (amplitude))
 
 static float *TableForFunc( genFunc_t func )
@@ -317,7 +318,7 @@ static void DeformText( const char *text ) {
 
 	// draw each character
 	for ( i = 0 ; i < len ; i++ ) {
-		ch = text[i];
+		ch = (byte)text[i];
 		ch &= 255;
 
 		if ( ch != ' ' ) {
@@ -742,6 +743,7 @@ void RB_CalcModulateColorsByFog( unsigned char *colors ) {
 	RB_CalcFogTexCoords( texCoords[0] );
 
 	for ( i = 0; i < tess.numVertexes; i++, colors += 4 ) {
+		// NOLINTNEXTLINE(clang-analyzer-core.CallAndMessage) — RB_CalcFogTexCoords initializes texCoords
 		float f = 1.0 - R_FogFactor( texCoords[i][0], texCoords[i][1] );
 		colors[0] *= f;
 		colors[1] *= f;
@@ -763,6 +765,7 @@ void RB_CalcModulateAlphasByFog( unsigned char *colors ) {
 	RB_CalcFogTexCoords( texCoords[0] );
 
 	for ( i = 0; i < tess.numVertexes; i++, colors += 4 ) {
+		// NOLINTNEXTLINE(clang-analyzer-core.CallAndMessage) — RB_CalcFogTexCoords initializes texCoords
 		float f = 1.0 - R_FogFactor( texCoords[i][0], texCoords[i][1] );
 		colors[3] *= f;
 	}
@@ -782,6 +785,7 @@ void RB_CalcModulateRGBAsByFog( unsigned char *colors ) {
 	RB_CalcFogTexCoords( texCoords[0] );
 
 	for ( i = 0; i < tess.numVertexes; i++, colors += 4 ) {
+		// NOLINTNEXTLINE(clang-analyzer-core.CallAndMessage) — RB_CalcFogTexCoords initializes texCoords
 		float f = 1.0 - R_FogFactor( texCoords[i][0], texCoords[i][1] );
 		colors[0] *= f;
 		colors[1] *= f;
